refactored model names to be singular
